Do not run inside docker
This turns the action into a Javascript to prevent using Docker (which
mostly adds overhead).

As documented by Github, this includes the node_modules directory with
the needed javascript modules. This also adds a minimal package.json,
but without a meaningful version number since this is just to track
dependencies.

This fixes #12.