Configure Arduino CLI to export binaries to the sketch folder
Arduino CLI 0.11.0 - 0.13.0 saved the compiled binaries to the build subfolder of the sketch by default. This behavior
will change in the next Arduino CLI release, so that the default is to not save the binaries.

The previous behavior may be most easily preserved in a backwards compatible manner by configuring Arduino CLI's
sketch.always_export_binaries configuration setting via an environment variable:
https://arduino.github.io/arduino-cli/latest/configuration/#environment-variables